Lesson 2 homework completed and ready for review

+++ src/main/java/students/diana_trekancenoka/lesson_2/level_3_junior/task_11/ArithmeticMean.java
@@ -6,11 +6,11 @@ public class ArithmeticMean {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter integer one: ");
-        int number = scanner.nextInt();
+        double number = scanner.nextInt();
         System.out.println("Enter integer two: ");
-        int number2 = scanner.nextInt();
+        double number2 = scanner.nextInt();
         System.out.println("Enter integer three: ");
-        int number3 = scanner.nextInt();
+        double number3 = scanner.nextInt();
 
         double result = (number + number2 + number3) / 3;
         System.out.println("Arithmetic mean is " + result);
